+++ packages/sh-design/src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './copy-button'
 export * from './lazy-image'
 export * from './seamless-scroll'
+export * from './waterfall'